fix: Rewrtie as a class with customizable web3 config
