+++ packages/pancake-profile-sdk/src/constants/common.ts
@@ -1,2 +1,5 @@
 export const profileApi = "https://profile.pancakeswap.com";
 export const profileSubgraphApi = "https://api.thegraph.com/subgraphs/name/pancakeswap/profile";
+
+export const MAINNET_CHAIN_ID = 56;
+export const TESTNET_CHAIN_ID = 97;
